fixed some stuff intellij pointed out

+++ dmoj/mnyc/MACKENZIE_NEW_YEARS_CHALLENGE_P2_MOORES_LAW.java
@@ -1,4 +1,6 @@
-import java.io.*;
+import java.io.BufferedReader;
+import java.io.IOException;
+import java.io.InputStreamReader;
 
 public class MACKENZIE_NEW_YEARS_CHALLENGE_P2_MOORES_LAW {
   public static void main (String [] args) throws IOException {
@@ -11,12 +13,12 @@ public class MACKENZIE_NEW_YEARS_CHALLENGE_P2_MOORES_LAW {
     TY = Double.parseDouble (String.format ("%.5f", TY * 365));
     
     Y = (int) (TY / 365);
-    TY -= (int)(Y * 365);
-    M = (int) (TY / 30);
-    TY -= (int)(M * 30);
-    W = (int) (TY / 7);
-    TY -= (int)(W * 7);
-    D = (int) Math.ceil (TY);
+	  TY -= Y * 365;
+	  M = (int) (TY / 30);
+	  TY -= M * 30;
+	  W = (int) (TY / 7);
+	  TY -= W * 7;
+	  D = (int) Math.ceil (TY);
     
     if (D >= 7) {
       D -= 7;
